Add filters for each category
