+++ packages/web/src/pages/search-page-v2/SearchPageV2.tsx
@@ -3,6 +3,7 @@ import { ChangeEvent, useCallback, useContext, useEffect } from 'react'
 import { Status } from '@audius/common/models'
 import { Maybe } from '@audius/common/utils'
 import {
+  FilterButton,
   Flex,
   IconAlbum,
   IconComponent,
@@ -30,102 +31,8 @@ import { useMedia } from 'hooks/useMedia'
 
 import { RecentSearches } from './RecentSearches'
 import { SearchCatalogTile } from './SearchCatalogTile'
-
-type Filter =
-  | 'genre'
-  | 'mood'
-  | 'key'
-  | 'bpm'
-  | 'isPremium'
-  | 'hasDownloads'
-  | 'isVerified'
-
-type Category = {
-  filters: Filter[]
-  icon?: IconComponent
-}
-
-const categories = {
-  all: { filters: [] },
-  profiles: { icon: IconUser, filters: ['genre', 'isVerified'] },
-  tracks: {
-    icon: IconNote,
-    filters: ['genre', 'mood', 'key', 'bpm', 'isPremium', 'hasDownloads']
-  },
-  albums: { icon: IconAlbum, filters: ['genre', 'mood'] },
-  playlists: { icon: IconPlaylists, filters: ['genre', 'mood'] }
-} satisfies Record<string, Category>
-
-type CategoryKey = keyof typeof categories
-
-type SearchHeaderProps = {
-  category?: CategoryKey
-  setCategory: (category: CategoryKey) => void
-  title: string
-  query: Maybe<string>
-}
-
-const SearchHeader = (props: SearchHeaderProps) => {
-  const { category: categoryKey = 'all', setCategory, query, title } = props
-
-  const { isMobile } = useMedia()
-
-  const handleChange = useCallback(
-    (e: ChangeEvent<HTMLInputElement>) => {
-      const value = e.target.value
-      setCategory(value as CategoryKey)
-    },
-    [setCategory]
-  )
-
-  const categoryRadioGroup = (
-    <RadioGroup
-      direction='row'
-      gap='s'
-      aria-label={'Select search category'}
-      name='searchcategory'
-      value={categoryKey}
-      onChange={handleChange}
-    >
-      {Object.entries(categories)
-        .filter(([key]) => !isMobile || key !== 'all')
-        .map(([key, category]) => (
-          <SelectablePill
-            aria-label={`${key} search category`}
-            icon={(category as Category).icon}
-            key={key}
-            label={capitalize(key)}
-            size='large'
-            type='radio'
-            value={key}
-            checked={key === categoryKey}
-          />
-        ))}
-    </RadioGroup>
-  )
-
-  return isMobile ? (
-    <Flex p='s' css={{ overflow: 'scroll' }}>
-      {categoryRadioGroup}
-    </Flex>
-  ) : (
-    <Header
-      {...props}
-      primary={title}
-      secondary={
-        query ? (
-          <Flex ml='l'>
-            <Text variant='heading' strength='weak'>
-              &#8220;{query}&#8221;
-            </Text>
-          </Flex>
-        ) : null
-      }
-      rightDecorator={categoryRadioGroup}
-      variant='main'
-    />
-  )
-}
+import { Category } from './types'
+import { CategoryKey, SearchHeader } from './SearchHeader'
 
 export const SearchPageV2 = () => {
   const { isMobile } = useMedia()
