+++ packages/web/src/pages/search-page-v2/types.ts
@@ -1,0 +1,15 @@
+import { IconComponent } from '@audius/harmony'
+
+export type Filter =
+  | 'genre'
+  | 'mood'
+  | 'key'
+  | 'bpm'
+  | 'isPremium'
+  | 'hasDownloads'
+  | 'isVerified'
+
+export type Category = {
+  filters: Filter[]
+  icon?: IconComponent
+}
